requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
apturl==0.5.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brlapi==0.8.3
cached-property==1.5.2
cachetools==5.3.3
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0
fasteners==0.14.1
fastjsonschema==2.20.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
future==0.18.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.26.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.1
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.25.0
ipython_genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.0
keras-tuner==1.4.7
keyring==23.5.0
kiwisolver==1.4.5
kt-legacy==1.0.5
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.52
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.77
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.4
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
nvidia-nccl-cu12==2.22.3
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.4
packages==0.1.1
packaging==23.2
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.45
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==2.4.7
pypdf==4.2.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
redis==3.5.3
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
requests-cache==0.5.2
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
ssh-import-id==5.11
stack-data==0.6.3
streamlit==1.35.0
streamlit-chat==0.1.1
systemd-python==234
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xdg==5
xgboost==2.1.0
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
